Migrate SecurityProvider tests closer to the internal client
As part of #1142, we want to reduce the transitive dependencies of
oapi-codegen.

We can start by breaking some dependencies between packages that don't
make sense.
