+++ internal/test/client/client_test.go
@@ -3,7 +3,12 @@ package client
 import (
 	"testing"
 
+	"github.com/deepmap/oapi-codegen/pkg/securityprovider"
 	"github.com/stretchr/testify/assert"
+)
+
+var (
+	withTrailingSlash = "https://my-api.com/some-base-url/v1/"
 )
 
 func TestTemp(t *testing.T) {
@@ -41,4 +46,34 @@ func TestTemp(t *testing.T) {
 	assert.Equal(t, expectedURL, client2.Server)
 	assert.Equal(t, expectedURL, client3.Server)
 	assert.Equal(t, expectedURL, client4.Server)
+}
+
+func TestSecurityProviders(t *testing.T) {
+	bearer, err := securityprovider.NewSecurityProviderBearerToken("mytoken")
+	assert.NoError(t, err)
+	client1, err := NewClient(
+		withTrailingSlash,
+		WithRequestEditorFn(bearer.Intercept),
+	)
+	assert.NoError(t, err)
+
+	apiKey, err := securityprovider.NewSecurityProviderApiKey("cookie", "apikey", "mykey")
+	assert.NoError(t, err)
+	client2, err := NewClient(
+		withTrailingSlash,
+		WithRequestEditorFn(apiKey.Intercept),
+	)
+	assert.NoError(t, err)
+
+	basicAuth, err := securityprovider.NewSecurityProviderBasicAuth("username", "password")
+	assert.NoError(t, err)
+	client3, err := NewClient(
+		withTrailingSlash,
+		WithRequestEditorFn(basicAuth.Intercept),
+	)
+	assert.NoError(t, err)
+
+	assert.Equal(t, withTrailingSlash, client1.Server)
+	assert.Equal(t, withTrailingSlash, client2.Server)
+	assert.Equal(t, withTrailingSlash, client3.Server)
 }
